Manage gobject references with correct Go object.
Previous attempts did not use the correct Go object,
but a temporary one created just for ref management.
This resulted in the object being collected early and
the ref decremented to 0. And so the gobject was
freed early, causing crashes.

This implementation use the correct Go object, and
early testing looks promising.