Generate reference management in all NewFromC funcs.
Trying to do it in a single generic function that takes an
arg of type interface{} won't work. A finalizer for an instance
of such a type is never called.

So there needs to be a type-specific implementation of
reference management. And so might as well put it in the
NewFromC functions themselves.